fix(server): serialize amounts as strings across all endpoints (A5)
Same amount, three wire types: create/detail/update emitted float
(create: float(amount); detail: round(amount,2), float(amount_original);
update: round(display_amount,2)) while list/batch/budget/statistics
endpoints used Decimal strings. 0.1+0.2-class float drift reached the UI
through those endpoints, and the detail vs update endpoints disagreed on
the amountOriginal type.

Now: TransactionDetailResponse/TransactionCreateResult/
TransactionUpdateResult declare Decimal, and crud_service passes the
stored Decimals through untouched (no float(), no round()) — every money
field serializes losslessly as a string. Floats remain only where they
belong: chart-adjacent ratios/percentages (statistics.py), never money.

Client unaffected: TransactionModel.amount is already Decimal and
_parseAmount accepts both numbers and strings.

The old test locked the drift itself (update=float vs detail=str); it is
reversed to lock Decimal type + lossless string serialization. 249 tests
pass; ruff/mypy clean.

+++ server/app/schemas/transaction.py
@@ -401,9 +401,9 @@ class TransactionDetailResponse(BaseModel):
     id: str  # UUID as string
     user_uuid: str = Field(..., serialization_alias="userUuid")
     type: str
-    amount: float
-    amount_original: float | None = Field(None, serialization_alias="amountOriginal")
-    amount_base: float = Field(..., serialization_alias="amountBase")
+    amount: Decimal
+    amount_original: Decimal | None = Field(None, serialization_alias="amountOriginal")
+    amount_base: Decimal = Field(..., serialization_alias="amountBase")
     currency: str
     base_currency: str = Field(..., serialization_alias="baseCurrency")
     exchange_rate: str | None = Field(None, serialization_alias="exchangeRate")
@@ -456,7 +456,7 @@ class TransactionCreateResult(BaseModel):
 
     success: bool
     transaction_id: str
-    amount: float
+    amount: Decimal
     currency: str
     type: str
     category_key: str
@@ -483,9 +483,9 @@ class TransactionUpdateResult(BaseModel):
 
     success: bool
     transaction_id: str
-    amount: float
-    amount_original: float
-    amount_base: float
+    amount: Decimal
+    amount_original: Decimal
+    amount_base: Decimal
     currency: str
     base_currency: str = Field(..., serialization_alias="baseCurrency")
     type: str
